FastAPI/FastAPI_App-classification/backend/requirements.txt
fastapi
uvicorn
torch
torchvision
pydantic
python-multipart
Pillow
aiofiles